Changed default number of generated lines to 1
This was a shortcut for running the program multiple times to get an
"agreeable" passphrase, which was a side-effect of the unfortunate
default Diceware word list.

Now that I've switched the default to the Diceware-Improved word list,
it's much more feasible to generate a single passphrase at a time and
get a "good" one.